requirements.txt
ansible==8.7.0
ansible-core==2.15.13
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==4.2.16
djangorestframework==3.15.2
gunicorn==23.0.0
importlib-resources==5.0.7
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.2
resolvelib==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
